refactor(buddies): extract BuddyRoleRepository to keep buddy_repository under size limit (#395)
